Add image re-sizer to web app
Image re-sizer functionality has been added to the web app to allow the resizing of images with dimensions greater than 1000 x 1000 pixels.

If a user image is larger than 1000 x 1000 pixels, it will get re-sized to 1000 x 1000.

This serves as a "fail-save"  in case users try to dreamify images that are too big. If an image is too big (3000 + pixels width or height)

The back-end service processing the image may fail due to lack of memory among other yet to be debugged reasons when an image is too big.